[MINOR] Use parse_version instead of LooseVersion to properly compare versions
When comparing the new version number to the previous version number to ensure that the version is being increased, the comparison was not working properly when going from a pre-release version to a release version, because `LooseVersion` is flawed and claims that `1.0.0-rc1` is greater than `1.0.0`. With this change, we use `parse_version` when comparing versions, so that `1.0.0` is treated as greater than `1.0.0-rc1`. We still use `LooseVersion` when breaking a version into its components, because `parse_version` does not support this.